+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-# ATUALIZAÇÕES DO PROJETO - Versão [1.1]
+# ATUALIZAÇÕES DO PROJETO - Versão [1.2]
 
 ## [1.1] - Atualização 12/12/2024:
 - Criar novos usuários
 - Informações de cada usuário são salvas em __arrays__
+
+## [1.2] - Atualização 01/01/2025:
+- Máximo de 3 tentativas ao digitar senha
+- Bloquear conta após 3 tentativas